requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
cryptography==45.0.5
openai==2.5.0
pytest==8.4.2
